LPS-59553 Rename
